[5.3] Language Association Codes
In the Language Code plugin it is possible to change the code used for a specified language. eg fr-FR can be changed to fr-CA

However everywhhere in the admin that associations are displayed it uses the original code and not the changed code.

This PR is a "takeover" of the original PR #37289 created by @infograf768

That PR has merge conflicts and @infograf768 is now retired.

See the original PR for more screenshots but an example is below

Signed-off-by: BrianTeeman <brian@teeman.net>