[5.3] RTL Calendar Days
When the calendar is being displayed with an RTL language the data in the calendar is correctly ordered RTL. However the names od the days is incorrect as this remains LTR

This PR adds a conditional check for RTL and reverses the string of days so that it is displayed correctly

Note that JS is not my thing and it looks like there is code for this already at line 140 but it is not working? Maybe a js expert can see why.

Signed-off-by: BrianTeeman <brian@teeman.net>