Unpublish/Run transition article when saving as copy
Second attempt at fixing https://github.com/joomla/joomla-cms/issues/37351.

Previous PR (https://github.com/joomla/joomla-cms/pull/38810) was reverted (in https://github.com/joomla/joomla-cms/pull/41790) because it would not work if Workflow was enabled.

This super simple PR checks if workflows are enabled and if so set a transition value of 1. If workflows not enabled, it sets state to 0, as in https://github.com/joomla/joomla-cms/pull/38810.