[53][Cypress] Don't Force SSL
With #44850 during Joomla installation the configuration parameter `force_ssl` is set to 2 (Force HTTP Entire Site).
This prevents the Joomla System Tests from running with HTTP. And as the test shows it is not necessary for the Joomla System Tests.
Even without explicit setting the parameter `force_ssl` the Joomla System Tests can be run over either HTTP or HTTPS.