[5.3] Author column
There is a check in the Author column to display the author name if the id is not 0 else to display the author as None

I believe this check to be incorrect. When a user is deleted from the users table the created_by in the content table is untouched so the condition in the query is never passed.

This PR changes the condition to check for the author_name and this way we can satisfy the condition and are able to display the author as None

This PR does not fix the related problem in the authors filter where you can not filter by author None even though it is intended that you can.

Signed-off-by: BrianTeeman <brian@teeman.net>