[5.3] Featured Author Column
### Summary of Changes
There is a check in the Author column to display the author name if the id is not 0 else to display the author as None

I believe this check to be incorrect. When a user is deleted from the users table the created_by in the content table is untouched so the condition in the query is never passed.

This PR changes the condition to check for the author_name in the featured articles and this way we can satisfy the condition and are able to display the author as None

_This PR is the same as #45170 but for the featured articles view_

### Testing Instructions
Create two users
Create featured articles with each user
Delete one of the users

Signed-off-by: BrianTeeman <brian@teeman.net>