Fix deprecation text and add Registry to StateBehaviorTrait
The `$state` property defined in the StateBehaviorTrait has as type hint the `State` class which is deprecated. The goal is to replace the `State` object with a `Registry` object. This can be done in 5.0+ by initialising the `$state` in the constructor with a `Registry` object. The Trait will use this `Registry` object. If the `$state` property is not type hinted with the Registry object, phpstan will complain about incorrect use of the `$state`.

This PR allows to use the correct type (Registry) and solves warnings from the static code analyser.