[6.0] Reset stats on update to 6
The data gathered by the Joomla stats plugin is hopelessly useless. If a site owner chose the option to send the stats "just once" then we never receive any stats ever again. The option to send the stats "just once" might have been made with Joomla 3.5 back in 2013 and we have never received any data for that site since. The site admin might have changed numerous times since then and they wont even know about the stats plugin.

All of this makes the data we have not fit for purpose.

This PR resets the params for the plugin so that any site updating to joomla 6 will be asked again if they are willing to help the Joomla project by submitting their site stats.

_personally I would make this reset happen on every minor release (approx every 6 months) but to be less controversial this PR makes the reset happen on the next major release only.